chore(team): split mailbox store relay paths
